{{ cookiecutter.bot_id.replace(' ', '_') }}/{{ cookiecutter.bot_id.replace(' ', '_') }}/custombot.py
from botcity.base import BaseBot
# Uncomment the line below for integrations with BotMaestro
# Using the Maestro SDK
# from botcity.maestro import *


class Bot(BaseBot):
    def action(self, execution=None):
        # Fetch the Activity ID from the task:
        # task = self.maestro.get_task(execution.task_id)
        # activity_id = task.activity_id

        ...

        # Uncomment to mark this task as finished on BotMaestro
        # self.maestro.finish_task(
        #     task_id=execution.task_id,
        #     status=AutomationTaskFinishStatus.SUCCESS,
        #     message="Task Finished OK."
        # )


if __name__ == '__main__':
    Bot.main()
